music_service/requirements.txt
alembic==1.13.2
aiobotocore
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
argcomplete==3.4.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.3
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.8
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
flower==2.0.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
humanize==4.10.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.3.7
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.6
packaging==24.1
passlib==1.7.4
pipx==1.6.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
pycparser==2.22
pydantic==2.8.2
pydantic-extra-types==2.9.0
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
timedelta==2020.12.3
tornado==6.4.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
userpath==1.9.2
uvicorn==0.30.1
uvloop==0.19.0
vine==5.1.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
gunicorn
loguru